refactor: Move SimpleTooltip component to a shared components directory
